Move phases fixtures to separate file
